WIP: enabled gaussian filtering of vector images
Changed the input pixel type list for the RecursiveGaussianImageFilter
to include VectorImages. The problem became that the output pixel type
needed to be of float components. ImageTypeTraits needed to be created
to work with both VectorImages and Images. However, the output is only
float doubles are down converted to floats.