WIP WIP WIP hacking at image registration method
WARNING: This topic will be re-based and hacked along the next
branch.

This is an initial prototype of what the ImageRegionstrationMethod may
look like in SimpleITK. It contains some examples to play with but no
tests. It's will dramatically change before being integrated. As many
Optimizers and Metrics have been added without testing to explore how
the interface and implementation will need to be done to be as
flexible as possible.

Change-Id: I37a198d959a3ce97ba8d23ae8c8189cfa4a74d9b