WIP: Add method to interface with SimpleITK Python Library
This is an initial prototype of a public method added to the SimpleITK
Python library ( _SimpleITK.so ) via Swig which allows the conversion
of a SimpleITK PyObject wrapped via Swig into a raw C++ pointer.

This should allow the C++ extension of methods for SimpleITK in Python
by providing the needed methods to convert from PyObject to SimpleITK
object ( or ITK objects ).

The interface of what objects to support and which direction still
needs to be done based on requirements, and if both ITK and SimpleITK
objects are needed.

Change-Id: I8ae293091471e0a30c24a82bf801c7592de4f662