WIP: Add scalar op vector operations.
Add support for multiplying a vector image by a constant scalar. This
patch changes the itk filter instantiated for all filters which use
the BinaryFunctorFilter template, from just the same binary type
for both images and scalars, to using a "ValueType" scalar image and
image type. This is the same as the pixel type for scalar images. For
VectorImages it is the same as the component type and for complex
images it is the same as the value_type.

WIP:  This has caused some problems for the complex pixel type with
the pow and add image filters. Specifically, the Add functor type use
the NumericTraits::AccumulatorType for std::complex<float> this is
std::complex<double>, but complex<double>+complex<float> is not
supported!!!

WIP: This now instantiates the BinaryOperator filters with the
combination of ScalarImages and VectorImages. The creates bigger
object files and libraries. But it does not exposed being able to
multiply scalar images with vector images.

Restores features that were provided by the MultipleByConstanst and
similarly named filters, before they were deprecated and removed.

Change-Id: I68143cf2a03d160136c030d5c7781d1f4db47308